Consolidate all Search Items into one module kodion.items.search_items
